Leet 50 110 226 review commit

+++ JavaCode/JavaCode/src/main/java/leet/methods/_9_DivideConquer/review/L50/Leet50_r1.java
@@ -4,7 +4,7 @@ public class Leet50_r1 {
 
     public double myPow(double x, int n) {
         if(n<0){
-            return mp(1/x,-n);
+            return mp(1/x,-(long) n);
         }
         return mp(x,n);
     }
